slider finish and indexBoxs , page1,2,3 =  total 60 results data fetch
