+++ src/atoms.ts
@@ -1,1 +1,7 @@
 import { atom } from "recoil";
+import { IMovie } from "./api";
+
+export const nowMovies = atom<IMovie[]>({
+  key: "nowMovies",
+  default: [],
+});
